Refactorisation en GO

+++ server/types/dio-resource.go
@@ -2,8 +2,8 @@ package types
 
 // DioResource : dio resource struct
 type DioResource struct {
-	Pin        int  `json:"pin"`
-	Sender     int  `json:"sender"`
-	Interuptor int  `json:"interruptor"`
-	On         bool `json:"on"`
+	Pin        int    `json:"pin"`
+	Sender     uint64 `json:"sender"`
+	Interuptor uint64 `json:"interruptor"`
+	On         bool   `json:"on"`
 }
